requirements.txt
flask
groq
pinecone-client
python-dotenv
pandas
scikit-learn
nltk
transformers
waitress